Deleting user: delete students from classes and avoiding teacher to delete their account if they have any class,

+++ app/src/main/java/com/example/teachly/EditProfile.java
@@ -32,7 +32,7 @@ public class EditProfile extends AppCompatActivity {
     TextView btnDeleteAccount, btnSaveAccount;
     private FirebaseAuth mAuth;
     EditText edtFullName, edtPhone, edtEmail;
-    String fullName, phone, email, uId;
+    String fullName, phone, email, uId, typeUser;
 
     FirebaseDatabase database;
     DatabaseReference referenceUsers, referenceClasses;
@@ -55,6 +55,7 @@ public class EditProfile extends AppCompatActivity {
         referenceClasses = database.getReference("classes");
         sharedPreferences = getSharedPreferences("Teachly", Context.MODE_PRIVATE);
         uId = sharedPreferences.getString("uId", "");
+        typeUser = getIntent().getStringExtra("typeUser");
 
         FirebaseUser currentUser = mAuth.getCurrentUser();
 
@@ -72,6 +73,58 @@ public class EditProfile extends AppCompatActivity {
                 builder.setPositiveButton("Delete Account", new DialogInterface.OnClickListener() {
                     @Override
                     public void onClick(DialogInterface dialog, int which) {
+
+                        if (typeUser.equals("Student")) {
+                            deleteStudent(currentUser);
+                        } else {
+                            deleteTeacher(currentUser);
+                        }
+                    }
+                });
+                builder.setIcon(android.R.drawable.ic_delete);
+                builder.setNeutralButton("Cancel", new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialog, int which) {
+                        dialog.dismiss();
+                    }
+                });
+                builder.setCancelable(false);
+                AlertDialog dialog = builder.create();
+                dialog.show();
+            }
+        });
+
+        btnSaveAccount.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                if (isNameChanged()) {
+                    Toast.makeText(EditProfile.this, "Full Name was updated!", Toast.LENGTH_SHORT).show();
+                }
+                if (isPhoneChanged()) {
+                    Toast.makeText(EditProfile.this, "Phone was updated!", Toast.LENGTH_SHORT).show();
+                }
+            }
+        });
+    }
+
+    private void deleteTeacher(FirebaseUser currentUser) {
+        DatabaseReference referenceClasses = FirebaseDatabase.getInstance().getReference("classes");
+
+        referenceClasses.addListenerForSingleValueEvent(new ValueEventListener() {
+            @Override
+            public void onDataChange(@NonNull DataSnapshot snapshot) {
+                Boolean teacherHasClass = false;
+                if (snapshot.exists()) {
+                    for (DataSnapshot snapshotClasses : snapshot.getChildren()) {
+                        String teacher = snapshotClasses.child("teacherUId").getValue(String.class);
+                        if (teacher.equals(uId)) {
+                            teacherHasClass = true;
+                        }
+                    }
+
+                    if (teacherHasClass) {
+                        Toast.makeText(EditProfile.this, "Delete ALL classes before removing your account!", Toast.LENGTH_SHORT).show();
+                    } else {
                         currentUser.delete().addOnCompleteListener(new OnCompleteListener<Void>() {
                             @Override
                             public void onComplete(@NonNull Task<Void> task) {
@@ -85,7 +138,6 @@ public class EditProfile extends AppCompatActivity {
                                                 snapshot.child(uId).getRef().removeValue();
 
                                                 removeFromCometChat(uId);
-                                                removeFromClass(uId);
 
                                                 SharedPreferences.Editor editor = sharedPreferences.edit();
                                                 editor.remove("uId");
@@ -122,28 +174,63 @@ public class EditProfile extends AppCompatActivity {
                             }
                         });
                     }
-                });
-                builder.setIcon(android.R.drawable.ic_delete);
-                builder.setNeutralButton("Cancel", new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialog, int which) {
-                        dialog.dismiss();
-                    }
-                });
-                builder.setCancelable(false);
-                AlertDialog dialog = builder.create();
-                dialog.show();
+                }
+            }
+
+            @Override
+            public void onCancelled(@NonNull DatabaseError error) {
+
             }
         });
+    }
 
-        btnSaveAccount.setOnClickListener(new View.OnClickListener() {
+    private void deleteStudent(FirebaseUser currentUser) {
+        currentUser.delete().addOnCompleteListener(new OnCompleteListener<Void>() {
             @Override
-            public void onClick(View v) {
-                if (isNameChanged()) {
-                    Toast.makeText(EditProfile.this, "Full Name was updated!", Toast.LENGTH_SHORT).show();
-                }
-                if (isPhoneChanged()) {
-                    Toast.makeText(EditProfile.this, "Phone was updated!", Toast.LENGTH_SHORT).show();
+            public void onComplete(@NonNull Task<Void> task) {
+                if (task.isSuccessful()){
+                    Query findUser = referenceUsers.orderByChild("userId").equalTo(uId);
+
+                    findUser.addListenerForSingleValueEvent(new ValueEventListener() {
+                        @Override
+                        public void onDataChange(@NonNull DataSnapshot snapshot) {
+                            if (snapshot.exists()) {
+                                snapshot.child(uId).getRef().removeValue();
+
+                                removeFromCometChat(uId);
+                                removeFromClass(uId);
+
+                                SharedPreferences.Editor editor = sharedPreferences.edit();
+                                editor.remove("uId");
+                                editor.remove("type");
+                                editor.apply();
+
+                                AlertDialog.Builder builder = new AlertDialog.Builder(EditProfile.this);
+                                builder.setTitle("Account deleted");
+                                builder.setMessage("Thanks for using Teachly!\nCome back soon!");
+
+                                builder.setPositiveButton("Ok", new DialogInterface.OnClickListener() {
+                                    @Override
+                                    public void onClick(DialogInterface dialog, int which) {
+                                        Intent intent = new Intent(EditProfile.this, Login.class);
+                                        startActivity(intent);
+                                        finish();
+                                        dialog.dismiss();
+                                    }
+                                });
+                                builder.setIcon(R.drawable.icon_check_30);
+                                builder.setCancelable(false);
+                                AlertDialog dialog = builder.create();
+                                dialog.show();
+                            }
+                        }
+                        @Override
+                        public void onCancelled(@NonNull DatabaseError error) {
+
+                        }
+                    });
+                } else {
+                    Toast.makeText(EditProfile.this, "Failed", Toast.LENGTH_SHORT).show();
                 }
             }
         });
@@ -201,7 +288,27 @@ public class EditProfile extends AppCompatActivity {
 
 
     private void removeFromClass(String uId){
+        DatabaseReference referenceClasses = FirebaseDatabase.getInstance().getReference("classes");
+        referenceClasses.addListenerForSingleValueEvent(new ValueEventListener() {
+            @Override
+            public void onDataChange(@NonNull DataSnapshot snapshot) {
+                if (snapshot.exists()) {
+                    for (DataSnapshot dataSnapshot : snapshot.getChildren())
+                    {
+                        DataSnapshot snapshotStudents = dataSnapshot.child("students");
+                        for (DataSnapshot student : snapshotStudents.getChildren()) {
+                            if (student.getValue(String.class).equals(uId)){
+                                student.getRef().removeValue();
+                            }
+                        }
+                    }
+                }
+            }
 
+            @Override
+            public void onCancelled(@NonNull DatabaseError error) {
 
+            }
+        });
     }
 }
